Client auth check flag used on MacOS
