+++ nym-vpn-core/crates/nym-ipc/src/authentication/mod.rs
@@ -11,7 +11,7 @@ mod macos;
 #[cfg(target_os = "macos")]
 pub use macos::SigningRequirements;
 #[cfg(target_os = "macos")]
-pub(crate) use macos::{Transport, incoming, is_authenticated};
+pub(crate) use macos::{Transport, incoming, is_authenticated, skip_authentication_checks};
 
 #[cfg(target_os = "windows")]
 mod windows;
